requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.3.0
Django==3.2.19
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.10.0
django-timezone-field==5.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.7.0
joblib==1.2.0
kombu==5.2.4
nltk==3.8.1
pip==23.1.2
prompt-toolkit==3.0.38
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
regex==2023.5.5
setuptools==67.7.2
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wheel==0.40.0
